+++ ksa_compliance/hooks.py
@@ -137,6 +137,11 @@ doc_events = {
         "validate": "ksa_compliance.standard_doctypes.sales_invoice.validate_sales_invoice",
         "before_cancel": "ksa_compliance.standard_doctypes.sales_invoice.prevent_cancellation_of_sales_invoice"
     },
+    "POS Invoice": {
+        "on_submit": "ksa_compliance.standard_doctypes.sales_invoice.create_sales_invoice_additional_fields_doctype",
+        "validate": "ksa_compliance.standard_doctypes.sales_invoice.validate_sales_invoice",
+        "before_cancel": "ksa_compliance.standard_doctypes.sales_invoice.prevent_cancellation_of_sales_invoice",
+    }
 }
 
 # Scheduled Tasks